+++ sources/fifteen/src/main/java/ru/bazalikova/fifteen/di/FifteenModule.kt
@@ -1,8 +1,9 @@
 package ru.bazalikova.fifteen.di
 
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
-import ru.bazalikova.fifteen.data.Game
+import ru.bazalikova.fifteen.domain.*
 import ru.bazalikova.fifteen.presentation.FifteenPresenter
 
 @Module
@@ -11,8 +12,17 @@ abstract class FifteenModule {
     companion object{
         @Provides
         @FifteenScope
-        fun providePresenter(game: Game): FifteenPresenter {
-            return FifteenPresenter(game)
+        fun providePresenter(getRandomFieldUseCase: GetRandomFieldUseCase, moveToCellUseCase: MoveToCellUseCase, checkGameOverUseCase: CheckGameOverUseCase): FifteenPresenter {
+            return FifteenPresenter(getRandomFieldUseCase, moveToCellUseCase, checkGameOverUseCase)
         }
     }
+
+    @Binds
+    abstract fun bindGetRandomUseCase(getRandomFieldUseCase: GetRandomFieldMainUseCase): GetRandomFieldUseCase
+
+    @Binds
+    abstract fun bindMoveToCellUseCase(moveToCellUseCase: MoveToCellMainUseCase): MoveToCellUseCase
+
+    @Binds
+    abstract fun bindCheckGameOverUseCase(checkGameOverUseCase: CheckGameOverMainUseCase): CheckGameOverUseCase
 }
