+++ sources/fifteen/src/main/java/ru/bazalikova/fifteen/domain/CheckGameOverUseCase.kt
@@ -1,0 +1,7 @@
+package ru.bazalikova.fifteen.domain
+
+import ru.bazalikova.fifteen.data.FifteenField
+
+interface CheckGameOverUseCase {
+    fun invoke(field: FifteenField) : Boolean
+}
