+++ sources/fifteen/src/main/java/ru/bazalikova/fifteen/domain/MoveToCellUseCase.kt
@@ -1,0 +1,7 @@
+package ru.bazalikova.fifteen.domain
+
+import ru.bazalikova.fifteen.data.FifteenField
+
+interface MoveToCellUseCase {
+    fun invoke(field: FifteenField, x: Int, y: Int) : Boolean
+}
