+++ sources/fifteen/src/main/java/ru/bazalikova/fifteen/presentation/FifteenPresenter.kt
@@ -2,12 +2,25 @@ package ru.bazalikova.fifteen.presentation
 
 import ru.bazalikova.fifteen.IFifteenMediator
 import ru.bazalikova.fifteen.data.Cell
-import ru.bazalikova.fifteen.data.Game
+import ru.bazalikova.fifteen.data.FifteenField
+import ru.bazalikova.fifteen.domain.CheckGameOverUseCase
+import ru.bazalikova.fifteen.domain.GetRandomFieldUseCase
+import ru.bazalikova.fifteen.domain.MoveToCellUseCase
 import javax.inject.Inject
 
-class FifteenPresenter @Inject constructor(private val game: Game) {
+class FifteenPresenter @Inject constructor(
+    val getRandomFieldUseCase: GetRandomFieldUseCase,
+    val moveToCellUseCase: MoveToCellUseCase,
+    val checkGameOverUseCase: CheckGameOverUseCase
+) {
     private var view: IFifteenView? = null
     private var mediator: IFifteenMediator? = null
+
+    companion object{
+        const val fieldCount = 3
+    }
+
+    private var field: FifteenField? = null
 
     fun attachView(view: IFifteenView)
     {
@@ -30,8 +43,10 @@ class FifteenPresenter @Inject constructor(private val game: Game) {
     }
 
     fun onViewCreated() {
-        val field = game.build()
+        val field = getRandomFieldUseCase.invoke(fieldCount)
         view?.setField(field)
+
+        this.field = field
     }
 
     fun onStart() {
@@ -43,12 +58,14 @@ class FifteenPresenter @Inject constructor(private val game: Game) {
     }
 
     fun move(cell: Cell) {
-        if (game.move(cell.row, cell.column))
+        val field = this.field ?: return
+
+        if (moveToCellUseCase.invoke(field, cell.row, cell.column))
         {
-            view?.setField(game.field)
+            view?.setField(field)
         }
 
-        if (game.checkGameOver())
+        if (checkGameOverUseCase.invoke(field))
         {
             view?.setGameOver(true)
             view?.setRepeatButton(true)
@@ -59,8 +76,10 @@ class FifteenPresenter @Inject constructor(private val game: Game) {
         view?.setGameOver(false)
         view?.setRepeatButton(false)
 
-        val field = game.build()
+        val field = getRandomFieldUseCase.invoke(fieldCount)
         view?.setField(field)
+
+        this.field = field
     }
 
     fun onCancelButtonClick()
